+++ lib/isolator/callbacks.rb
@@ -11,6 +11,14 @@ module Isolator
       after_isolate_callbacks << block
     end
 
+    def on_transaction_begin(&block)
+      transaction_begin_callbacks << block
+    end
+
+    def on_transaction_end(&block)
+      transaction_end_callbacks << block
+    end
+
     def start!
       return if Isolator.disabled?
       before_isolate_callbacks.each(&:call)
@@ -21,12 +29,28 @@ module Isolator
       after_isolate_callbacks.each(&:call)
     end
 
+    def notify!(event, payload)
+      if event == :begin
+        transaction_begin_callbacks.each { |cb| cb.call(payload) }
+      elsif event == :end
+        transaction_end_callbacks.each { |cb| cb.call(payload) }
+      end
+    end
+
     def before_isolate_callbacks
       @before_isolate_callbacks ||= []
     end
 
     def after_isolate_callbacks
       @after_isolate_callbacks ||= []
+    end
+
+    def transaction_begin_callbacks
+      @transaction_begin_callbacks ||= []
+    end
+
+    def transaction_end_callbacks
+      @transaction_end_callbacks ||= []
     end
   end
 end